Se mejora el manejo de validaciones haciendo que salgan todos los errores en una solo mensaje. Se implementan requerimientos 2,3 y 4 de Book

+++ src/main/java/com/accenture/demobookify/model/Book.java
@@ -37,7 +37,7 @@ public class Book {
         this.author = author;
         this.description = datosBook.description();
         this.price = datosBook.price();
-        this.quantity_available = datosBook.quantity_available();
+        this.quantity_available = (int) datosBook.quantity_available();
         this.isActive = true;
     }
